// types/ChartType.type.ts
export enum ChartType {
  Services = 'Services',
  Income = 'Income',
  ClientsNewReturn = 'ClientsNewReturn',
  Visits = 'Visits',
  Messages = 'Messages',
  TreatedClients = 'TreatedClients',
  DoctorsIncome = 'DoctorsIncome',
  DoctorsVisits = 'DoctorsVisits',
  DoctorsConversion = 'DoctorsConversion',
  ClientsSource = 'ClientsSource',
}